Change lens from double to int to allow zoom by changing the lens value instead of the camera position

+++ gama.core/src/gama/core/outputs/LayeredDisplayData.java
@@ -866,7 +866,7 @@ public class LayeredDisplayData {
 	/**
 	 * @return the cameraLens
 	 */
-	public int getCameraLens() { return camera.getLens(); }
+	public double getCameraLens() { return camera.getLens(); }
 
 	/**
 	 * Gets the preset camera.
